pyproject.toml
[project]
name = "Sakiko"
version = "0.1.0"
description = "基于 NoneBot2 和 LiteLoaderBDS 的 Minecraft 基岩版 Bot"
authors = [
    {name = "zhaomaoniu", email = "2667292003@qq.com"},
]
dependencies = [
    "nonebot2[aiohttp]>=2.1.2",
    "nonebot-plugin-alconna>=0.33.9",
    "nonebot-plugin-userinfo>=0.1.3",
    "nonebot-adapter-llbds>=0.1.1",
    "llpy-helper-lib>=2.13.1",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

